+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Flash messages failed in view helper script (https://github.com/PierreRambaud/GotCms/issues/28)
 - Update to Zend Framework 2.1.3
 - Use new GotSniffs based on PSR2 (https://github.com/PierreRambaud/GotCms/issues/38)
+- Update jquery, jquery ui and modernizr (https://github.com/PierreRambaud/GotCms/issues/39)
 
 ## 0.1.4 (12 Feb 2013):
 - Bug with HTTP 404 (undefined variable view)